.history/requirements_20251029134405.txt
# App requirements for Streamlit Community Cloud
streamlit>=1.37.0
torch
torchvision
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
# If you need OpenCV for heatmap.py locally:
# opencv-python
